+++ changelog.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [unreleased]
 
 - Add option --alias-prefix
-- Add flag --write-aliases
+- Add option --write-aliases
 
 ## [0.5.0] - 2022-04-12